adds named export reducers refactors createRootReducer from index closes #1862
