+++ src/lib/MiradorViewer.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { I18nextProvider } from 'react-i18next';
 import deepmerge from 'deepmerge';
 import App from '../containers/App';
-import createRootReducer from '../state/reducers/index';
+import createRootReducer from '../state/reducers/rootReducer';
 import createStore from '../state/createStore';
 import * as actions from '../state/actions';
 import settings from '../config/settings';
