+++ src/state/createStore.js
@@ -6,7 +6,7 @@
 import thunkMiddleware from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import createRootReducer from './reducers/index';
+import createRootReducer from './reducers/rootReducer';
 
 
 /**
