+++ src/state/reducers/index.js
@@ -1,22 +1,5 @@
-import { combineReducers } from 'redux';
-import workspaceReducer from './workspace';
-import windowsReducer from './windows';
-import manifestsReducer from './manifests';
-import infoResponsesReducer from './infoResponses';
-import configReducer from './config';
-
-/**
- * Function to create root reducer
- * from plugin reducers.
- * @namespace CreateRootReducer
- */
-export default function createRootReducer(pluginReducers) {
-  return combineReducers({
-    workspace: workspaceReducer,
-    windows: windowsReducer,
-    manifests: manifestsReducer,
-    infoResponses: infoResponsesReducer,
-    config: configReducer,
-    ...pluginReducers,
-  });
-}
+export * from './workspace';
+export * from './windows';
+export * from './manifests';
+export * from './infoResponses';
+export * from './config';
