+++ src/state/reducers/workspace.js
@@ -3,7 +3,7 @@ import ActionTypes from '../actions/action-types';
 /**
  * workspaceReducer
  */
-const workspaceReducer = (state = {}, action) => {
+export const workspaceReducer = (state = {}, action) => {
   switch (action.type) {
     case ActionTypes.FOCUS_WINDOW:
       return { ...state, focusedWindowId: action.windowId };
@@ -19,5 +19,3 @@ const workspaceReducer = (state = {}, action) => {
       return state;
   }
 };
-
-export { workspaceReducer as default };
